ml-processor/requirements.txt
ultralytics>=8.0.0
opencv-python>=4.7.0
fastapi>=0.100.0
pydantic>=2.0.0
uvicorn>=0.22.0
python-multipart>=0.0.6
pillow>=9.5.0
numpy>=1.24.0  # Often needed for array operations with images
torch>=2.0.0  # Required by ultralytics